feat: use LibreHardwareMonitor instead of OpenHardwareMonitor to get CPU and GPU temperatures
Replace the usage of OpenHardwareMonitorLib with LibreHardwareMonitorLib due to the former's lack of maintenance and the latter's support of newer and more hardware.

BREAKING CHANGE: Removes the boolean `average` argument in the `GPU_Temp` and `CPU_Temp` functions, and changes their return type to a float temperature instead of a list of temperatures (the behavior with `average=True` with the prior code).