fix: add a fix for __version__ not being included in the package once built
This commit adds a fix for __version not being included in the final package once it is build. This was caused by the package only importing __version__ if the _version file exists, though it did not check this in relation to the __init__.py file but rather the location of where it was ran.